+++ common/src/main/java/net/mehvahdjukaar/vista/client/renderer/LevelRendererCameraState.java
@@ -1,7 +1,6 @@
 package net.mehvahdjukaar.vista.client.renderer;
 
 import it.unimi.dsi.fastutil.objects.ObjectArrayList;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.LevelRenderer;
 import net.minecraft.client.renderer.SectionOcclusionGraph;
 import net.minecraft.client.renderer.ViewArea;
@@ -21,26 +20,16 @@ public class LevelRendererCameraState {
     @Nullable
     private ViewArea viewArea; //same as the actual one as this doesnt change actualy. unless we want to add it in the fufture to make far away cameras load
     private int lastViewDistance;
+    @Nullable
+    //lazy initialized
     private SectionOcclusionGraph sectionOcclusionGraph;
     private ObjectArrayList<SectionRenderDispatcher.RenderSection> visibleSections = new ObjectArrayList<>(10000);
 
-    private LevelRendererCameraState() {
-    }
-
-    public static LevelRendererCameraState createNew() {
-        var instance = new LevelRendererCameraState();
-        instance.sectionOcclusionGraph = new SectionOcclusionGraph();
-        Minecraft mc = Minecraft.getInstance();
-        LevelRenderer lr = mc.levelRenderer;
-     //   instance.viewArea = new ViewArea(lr.sectionRenderDispatcher, mc.level,
-                //TODO: change this
-      //          mc.options.getEffectiveRenderDistance(), lr);
-        instance.sectionOcclusionGraph.waitAndReset(instance.viewArea);
-        return instance;
+    public LevelRendererCameraState() {
     }
 
     public void copyFrom(LevelRenderer lr) {
-       // this.viewArea = lr.viewArea;
+        // this.viewArea = lr.viewArea;
         this.lastViewDistance = lr.lastViewDistance;
         this.sectionOcclusionGraph = lr.sectionOcclusionGraph;
         this.lastCameraSectionX = lr.lastCameraSectionX;
@@ -61,7 +50,13 @@ public class LevelRendererCameraState {
     }
 
     public void apply(LevelRenderer lr) {
-      //  lr.viewArea = this.viewArea;
+        //initialize graph
+        if (this.sectionOcclusionGraph == null) {
+            this.sectionOcclusionGraph = new SectionOcclusionGraph();
+            this.sectionOcclusionGraph.waitAndReset(lr.viewArea);
+        }
+
+        //  lr.viewArea = this.viewArea;
         lr.sectionOcclusionGraph = this.sectionOcclusionGraph;
         lr.visibleSections = this.visibleSections;
         lr.lastViewDistance = this.lastViewDistance;